% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract information from search results}
\usage{
extract(results, element = c("id", "caseNumber", "judges", "judgmentType",
  "textContent", "keywords", "division", "judgmentDate"))
}
\arguments{
\item{results}{list, result of searching}

\item{element}{Piece of information to extract}
}
\value{
data.frame with specific information
}
\description{
Extrac specific elements
}
\details{
Elements:
id
caseNumber
judges
judgmentType
textContent
keywords
division
judgmentDate
}

