% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_courts}
\alias{get_courts}
\title{Download all courts}
\usage{
get_courts()
}
\value{
data.frame with rows corresponding to courts and columns TO DO
}
\description{
Download information about all courts in the repository.
}
\seealso{
\code{\link[saos]{get_scChambers}}
}

