% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_courts.R
\name{get_dump_courts}
\alias{get_dump_courts}
\title{Download all common courts}
\usage{
get_dump_courts()
}
\value{
data.frame as described in \code{\link[saos]{courts}}
}
\description{
Download information about all common courts in the repository.
}
\seealso{
\code{\link[saos]{courts}}, \code{\link[saos]{get_scChambers}}
}

