% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_dump_scChambers.R
\name{get_dump_scChambers}
\alias{get_dump_scChambers}
\title{Download all Supreme Court chambers}
\usage{
get_dump_scChambers(simplify = FALSE)
}
\arguments{
\item{simplify}{Logical. If \code{TRUE} results will be returned as
  \code{data.frame}.}
}
\value{
The list of all Supreme Courts chambers as returned from API.
}
\description{
Download information about all chambers of the Supreme Court of Poland.
}
\seealso{
\code{\link[saos]{scchambers}}, \code{\link[saos]{get_dump_courts}}
}

