% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_judgments_texts.R
\name{get_judgment_texts}
\alias{get_judgment_texts}
\title{Get full texts of given judgments}
\usage{
get_judgment_texts(idlist)
}
\arguments{
\item{idlist}{integer vector with judgments' IDs}
}
\value{
named character vector with judgments
}
\description{
Get full texts, if available, of judgments with IDs from a given vector
}
\examples{
\donttest{
get_judgment_texts(128334)
get_judgment_texts(c(128334, 77354))
 }
}

