% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_judgments.R
\name{get_judgments}
\alias{get_judgments}
\alias{get_judgments.numeric}
\alias{get_judgments.saos_search}
\title{Get all data about judgments}
\usage{
get_judgments(x, verbose = TRUE)

\method{get_judgments}{numeric}(x, verbose = TRUE)

\method{get_judgments}{saos_search}(x, verbose = TRUE)
}
\arguments{
\item{x}{An object used to select a method.}

\item{verbose}{Logical. Whether or not display progress bar.}
}
\value{
A list with elements corresponding to judgments, as returned from API.
}
\description{
Get_judgments will download all information available about given judgments.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Download judgments, if available, with IDs from a
  given vector. If judgment is not available (meaning it doesn't exist in
  database) it will be skipped and warning will be generated.

\item \code{saos_search}: Get details of all judgments in given search
  results.
}}
\examples{
\donttest{
## From list of IDs
# single ID
judgments1 <- get_judgments(1)

# vector of IDs
judgments2 <- get_judgments(c(1, 100))

# vector of IDs with non-existent judgment
judgments3 <- get_judgments(c(128334, 1, 100))


## From search results
search <- search_judgments(limit = 10, verbose = FALSE)
judgments4 <- get_judgments(search)
 }
}

