% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_judgments_dump.R
\name{get_judgments_dump}
\alias{get_judgments_dump}
\title{Get all judgments}
\usage{
get_judgments_dump(start_date = NULL, end_date = NULL,
  modification_date = NULL)
}
\arguments{
\item{start_date}{Any date/time object that could be properly converted with
\code{as.POSIXct}, or a string in a format "\%Y-\%m-\%d".
Represents the earliest allowed judgment's date on the list.
If missing, no time limit is set.}

\item{end_date}{Any date/time object that could be properly converted with
\code{as.POSIXct}, or a string in a format "\%Y-\%m-\%d".
Represents the latest allowed judgment's date on the list.
If missing, no time limit is set.}

\item{modification_date}{Any date/time object that could be properly converted with
  \code{as.POSIXct}, or a string in format "\%Y-\%m-\%dT\%H:\%M:\%S".
  Allows to select judgments which were modified later than the specified
  time. If missing, no time limit is set.}
}
\value{
List of judgments as returned from API.
}
\description{
Get the list of all judgments within (and including) given dates.
}
\examples{
\dontrun{
full <- get_judgments_dump()
lastchanges <- get_judgments_dump(modification_date = Sys.Date() - 7)
}
\donttest{
# judgments from last week
lastweek <- get_judgments_dump(start_date = Sys.Date() - 7,
                               end_date = Sys.Date())
 }
}

