% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_scChambers}
\alias{get_scChambers}
\title{Download all Supreme Court chambers}
\usage{
get_scChambers(flatten = FALSE)
}
\arguments{
\item{flatten}{logical, whether flatten information about divisions of every
 chamber}
}
\value{
If \code{flatten = FALSE} (default) data.frame with rows
  corresponding to chambers and three columns
\tabular{rlll}{
n \tab name \tab type \tab description \cr
[,1] \tab id \tab integer \tab ID in the repository \cr
[,2] \tab name \tab character \tab full name of the chamber \cr
[,3] \tab divisions \tab list \tab details of the chamber's divisions, see
  details below \cr
}

Information about divisions is stored in dataframes with following columns.
\tabular{rlll}{
n \tab name \tab type \tab description \cr
[,1] \tab id \tab integer \tab uniue ID \cr
[,2] \tab name \tab character \tab name of the division \cr
[,3] \tab fullName \tab character \tab full name of a division, composed from
  names of both chamber and its division \cr
}

If \code{flatten = TRUE} data.frame with rows corresponding to divisions
  and five columns: information about chamber (\code{id, name}) and
  information about division (\code{div_id, div_name, div_fullName}).
}
\description{
Download information about all chambers of the Supreme Court of Poland.
}
\seealso{
\code{\link[saos]{get_courts}}
}

