% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/open_judgments.R
\name{open_judgments}
\alias{open_judgments}
\title{Open SAOS judgments' pages}
\usage{
open_judgments(judgments, limit = 5)
}
\arguments{
\item{judgments}{Data frame of search results, as returned from
\code{\link{search_judgments}} or \code{\link{get_judgments}}.}

\item{limit}{Limit the number of opened pages.}
}
\value{
(Invisibly) Vector of character strings with URLs of abstracts opened.
}
\description{
Open in web browser SAOS pages for given judgments.
}
\examples{
\donttest{
judgments <- search_judgments()
open_judgments(judgments)
 }
}

