% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search_judgments}
\alias{search_judgments}
\title{Search for judgments}
\usage{
search_judgments(query = NULL, limit = 200, force = FALSE)
}
\arguments{
\item{query}{Search pattern as a named list; pairs name-value are treated
as query parameters with corresponding values, see Details.}

\item{limit}{Limit the number of search results.}

\item{force}{If TRUE, force search request even if it seems extreme.}
}
\value{
data.frame with rows corresponding to judgments, or \code{NULL} if
 none judgment is available, and following columns (* means that column is
 a list with elements of given type, if impossible to unlist)
\tabular{rlll}{
n \tab name \tab type \tab description \cr
[,1] \tab id \tab integer \tab judgment ID in the repository \cr
[,2] \tab href \tab character \tab link to judgment's page in API \cr
[,3] \tab courtCases \tab character \tab case signature, indicates division
  of the trial court, a case number and a year of the judgment \cr
[,4] \tab judgmentType \tab character \tab one from "DECISION", "RESOLUTION",
  "SENTENCE", "REGULATION", "REASONS" \cr
[,5] \tab judges \tab data.frame* \tab two columns with judges names ("name")
  and their role ("specialRoles") \cr
[,6] \tab textContent \tab character \tab shortened version of full text of
  the judgment \cr
[,7] \tab keywords \tab character* \tab keywords associated with the given
  judgment \cr
[,8] \tab division \tab data.frame* \tab information about the court (columns
  "court.name", "court.code", "court.href") and its division (columns "name",
  "code", "href"), see \link{get_courts} for details \cr
[,9] \tab judgmentDate \tab character \tab date of the judgment in format
  YYYY-MM-DD \cr
}
}
\description{
Search for judgments matching the given query
}
\details{
The available query parameters: \cr
 \code{all} - character, search everywhere, \cr
 \code{legalBase} - character, \cr
 \code{referencedRegulation} - character, \cr
 \code{keyword} - character, exact matching, \cr
 \code{courtName} - character, exact matching, \cr
 \code{judgeName} - character, \cr
 \code{judgmentDateFrom} - character in format "YYYY-MM-DD", \cr
 \code{judgmentDateTo} - character in format "YYYY-MM-DD".
}
\examples{
\dontrun{
search_judgments(list(judgmentDateFrom = "2014-11-20"))
search_judgments(list(judgeName="Maria Tyszel", judgmentDateTo="2014-06-30"))

# search with no query, various limit options
search_judgments()
search_judgments(limit = 50)
search_judgments(limit = 300)
search_judgments(limit = 300, force = TRUE)
 }
}

