% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count_judgments}
\alias{count_judgments}
\title{Count number of results for a given search}
\usage{
count_judgments(query = NULL)
}
\arguments{
\item{query}{Search pattern as a named list; pairs name-value are treated
  as query parameters with corresponding values, see Details.}
}
\value{
the number of results (integer)
}
\description{
Count the number of results for a given search. Useful to check before
 attempting to download a huge number of records, which could be very
 memory and time consuming.
}
\details{
The available query parameters:
}
\examples{
\dontrun{
 count_judgments(list(dateFrom="01-11-2014"))
 count_judgments(list(judgeName="Maria+Tyszel", dateTo="01-11-1995"))
  }
}
\seealso{
\code{\link[saos]{search_judgments}}
}

