% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count_judgments}
\alias{count_judgments}
\title{Count number of results for a given search}
\usage{
count_judgments(query = NULL)
}
\arguments{
\item{query}{Search pattern as a named list; pairs name-value are treated
  as query parameters with corresponding values, see Details.}
}
\value{
the number of results (integer)
}
\description{
Count the number of results for a given search. Useful to check before
 attempting to download a huge number of records, which could be very
 memory and time consuming.
}
\details{
The available query parameters: \cr
 \code{all} - character, search everywhere, \cr
 \code{legalBase} - character, \cr
 \code{referencedRegulation} - character, \cr
 \code{keyword} - character, exact matching, \cr
 \code{courtName} - character, exact matching, \cr
 \code{judgeName} - character, \cr
 \code{judgmentDateFrom} - character in format "YYYY-MM-DD", \cr
 \code{judgmentDateTo} - character in format "YYYY-MM-DD".
}
\examples{
\dontrun{
 count_judgments(list(judgmentDateFrom="2014-01-01"))
 count_judgments(list(judgeName="Maria Tyszel", judgmentDateTo="2014-06-30"))
  }
}
\seealso{
\code{\link[saos]{search_judgments}}
}

