% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count_judgments}
\alias{count_judgments}
\title{Count number of results for a given search}
\usage{
count_judgments(all = NULL, legalBase = NULL, referencedRegulation = NULL,
  keyword = NULL, courtName = NULL, judgeName = NULL,
  judgmentDateFrom = NULL, judgmentDateTo = NULL)
}
\arguments{
\item{all}{search everywhere for given phrase}

\item{legalBase}{legal basis of judgments}

\item{referencedRegulation}{regulations referenced in judgments}

\item{keyword}{keywords of judgments}

\item{courtName}{name of the trial court}

\item{judgeName}{name of any judge}

\item{judgmentDateFrom}{judgments from and including given date,
format "YYYY-MM-DD"}

\item{judgmentDateTo}{judgments to and including given date,
format "YYYY-MM-DD"}
}
\value{
the number of results (integer)
}
\description{
Count the number of results for a given search. Useful to check before
 attempting to download a huge number of records, which could be very
 memory and time consuming.
}
\examples{
\donttest{
 count_judgments(judgmentDateFrom="2014-01-01")
 count_judgments(judgeName="Maria Tyszel", judgmentDateTo="2014-06-30")
  }
}
\seealso{
\code{\link[saos]{search_judgments}}
}

