% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract.saos_judgments}
\alias{extract.saos_judgments_dump}
\alias{extract.saos_search}
\title{Extract information from judgments' list}
\usage{
extract(x, element)

\method{extract}{saos_search}(x, element)

\method{extract}{saos_judgments}(x, element)

\method{extract}{saos_judgments_dump}(x, element)
}
\arguments{
\item{x}{An object used to select a method.}

\item{element}{Name of element to extract, see Details.}
}
\value{
A data.frame with first column indicating ID of judgment and the rest
  depending on which element was extracted, see Details.
}
\description{
Extract as a data frame specific elements from judgements' list as returned
  from \code{\link{search_judgments}}, \code{\link{get_judgments}} or
  \code{\link{get_dump_judgments}}.
}
\details{
Available elements for all methods are:
\itemize{
 \item id,
 \item courtCases,
 \item judgmentType,
 \item judges,
 \item textContent,
 \item keywords,
 \item division,
 \item judgmentDate.
 }
 Additionally for classes \code{saos_judgments} and \code{saos_judgments_dump}
   there are more elements:
 \itemize{
 \item courtType,
 \item source,
 \item courtReporters,
 \item decision,
 \item summary,
 \item legalBases,
 \item referencedRegulations,
 \item referencedCourtCases,
 \item personnelType,
 \item judgmentForm (\code{saos_judgments}) or form (\code{saos_judgments_dump}),
 \item chambers.
 }

 For class \code{saos_judgments} there is also element "href".

Data frame with data for all given judgments is always returned. If there is
  no data, a row with only \code{NA} (except for id) is insertes in the data
  frame. Data frames are always in long format, for example when one judge
  has a few roles, his name is multiplied accordingly.
}
\section{Methods (by class)}{
\itemize{
\item \code{saos_search}: Extract data from list of results from
  \code{\link{search_judgments}}.

\item \code{saos_judgments}: Extract data from list of judgments from
  \code{\link{get_judgments}}.

\item \code{saos_judgments_dump}: Extract data from list of judgments from
  \code{\link{get_dump_judgments}}.
}}
\examples{
\donttest{
saos_search <- search_judgments(limit = 50, verbose = FALSE)
saos_judgments <- get_judgments(saos_search, verbose = FALSE)

# for class "saos_search"
judges <- extract(saos_search, "judges")
dim(judges)
names(judges)

divisions <- extract(saos_search, "division")
dim(divisions)
names(divisions)


# for class "saos_judgments"
sources <- extract(saos_judgments, "source")
dim(sources)
names(sources)

ref_reg <- extract(saos_judgments, "referencedRegulations")
dim(ref_reg)
names(ref_reg)


## for class "saos_judgments_dump"
judgments <- get_dup_judgments(start_date = "2015-01-01",
                               end_date = "2015-10-01", verbose = FALSE)
court <- extract(judgments, "courtType")
dim(court)
names(court)
}
}

