% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_courts}
\alias{get_courts}
\title{Download all courts}
\usage{
get_courts()
}
\value{
data.frame with rows corresponding to courts and following columns
\tabular{rlll}{
n \tab name \tab type \tab description \cr
[,1] \tab id \tab integer \tab court ID in the repository \cr
[,2] \tab name \tab character \tab full name of the court \cr
[,3] \tab type \tab character \tab type of the court; district, regional or
  appeal \cr
[,4] \tab code \tab character \tab court's code, see details below \cr
[,5] \tab parentCourt \tab integer \tab ID of the superior court \cr
[,6] \tab divisions \tab list \tab details of the court's divisions, see
  details below   \cr
}
Court's code represents nested structure. It is in format "15BBCCDD" where
  "BB" - code of the coresponding appeal court,
  "CC" - code of the corresponding regional court ("00" for appeal courts),
  "DD" - code of the district court ("00" for appeal and regional courts).
 All codes are multiples of five.

Information about divisions is stored in dataframes with following columns.
\tabular{rlll}{
n \tab name \tab type \tab description \cr
[,1] \tab id \tab integer \tab unique ID \cr
[,2] \tab name \tab character \tab full name of the division \cr
[,3] \tab code \tab character \tab court's code, see details below \cr
[,4] \tab type \tab character \tab type of the division \cr
}
Code of the division is in the format "AABCCDD", where
  "AA" - usually  "00", or number of affiliate division,
  "B" - usually "0", in case of affiliate divisions "1" or "2",
  "CC" - number of division in the given court (multiples of five),
  "DD" - code corresponding to the type of the division.

Information in the repository should be thorough (no missing data).
}
\description{
Download information about all courts in the repository.
}
\seealso{
\code{\link[saos]{get_scChambers}}
}

