% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_dump_courts.R
\name{get_dump_courts}
\alias{get_dump_courts}
\title{Download all common courts}
\usage{
get_dump_courts(simplify = FALSE)
}
\arguments{
\item{simplify}{Logical. If \code{TRUE} results will be returned as
\code{data.frame}.}
}
\value{
The list of all common courts as returned from API.
}
\description{
Download information about all common courts in the repository.
}
\seealso{
\code{\link[saos]{courts}}, \code{\link[saos]{get_dump_scChambers}}
}

