% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_dump_courts.R
\encoding{UTF-8}
\name{get_dump_courts}
\alias{get_dump_courts}
\title{Download all common courts}
\usage{
get_dump_courts(simplify = FALSE)
}
\arguments{
\item{simplify}{Logical. If \code{TRUE} results will be returned as
\code{data.frame}.}
}
\value{
If \code{simplify = FALSE} the list of all common courts as returned
  from API. Every element of the list represents one court and has the
  following, illustrative structure: \cr
  List of 6 \cr
   $ id         : int 1 \cr
   $ name       : chr "Sąd Apelacyjny we Wrocławiu" \cr
   $ type       : chr "APPEAL" \cr
   $ code       : chr "15500000" \cr
   $ parentCourt: NULL \cr
   $ divisions  :List of 3 \cr
   ..$ :List of 4 \cr
   .. ..$ id  : int 1 \cr
   .. ..$ name: chr "I Wydział Cywilny" \cr
   .. ..$ code: chr "0000503" \cr
   .. ..$ type: chr "Cywilny" \cr
   ..$ :List of 4 \cr
   .. ..$ id  : int 2 \cr
   .. ..$ name: chr "II Wydział Karny" \cr
   .. ..$ code: chr "0001006" \cr
   .. ..$ type: chr "Karny" \cr
   ..$ :List of 4 \cr
   .. ..$ id  : int 3 \cr
   .. ..$ name: chr "III Wydział Pracy i Ubezpieczeń Społecznych" \cr
   .. ..$ code: chr "0001521" \cr
   .. ..$ type: chr "Pracy i Ubezpieczeń Społecznych" \cr

   Detailed description of the meaning of all elements could be found in
   \code{\link[saos]{courts}}.

   If \code{simplify = TRUE} a \code{data.frame} as described in
   \code{\link[saos]{courts}} is returned.
}
\description{
Download information about all common courts in the repository.
}
\seealso{
\code{\link[saos]{courts}}, \code{\link[saos]{get_dump_scChambers}}
}

