% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_dump_enrichments.R
\encoding{UTF-8}
\name{get_dump_enrichments}
\alias{get_dump_enrichments}
\title{Download all enrichments tags (EXPERIMENTAL)}
\usage{
get_dump_enrichments(simplify = FALSE)
}
\arguments{
\item{simplify}{Logical. If \code{TRUE} results will be returned as
\code{data.frame}.}
}
\value{
If \code{simplify = FALSE} the list of enrichment tags as returned
  from API. Every element of the list represents one tag and has the
  following, illustrative structure: \cr
 List of 4
 $ id        : int 204 \cr
 $ judgmentId: int 31345 \cr
 $ tagType   : chr "REFERENCED_CASE_NUMBERS" \cr
 $ value     :List of 2 \cr
 ..$ referencedCaseNumber: chr "III CK 430/03" \cr
 ..$ referencedIds       :List of 1 \cr
 .. ..$ : int 1728 \cr

 Detailed description of the meaning of all elements could be found below.

 If \code{simplify = TRUE} a \code{data.frame} with following columns:
\tabular{rlll}{
n \tab name \tab class \tab description \cr
[,1] \tab id \tab integer \tab ID of a tag \cr
[,2] \tab judgmentId \tab integer \tab ID of tagged judgment \cr
[,3] \tab tagType \tab character \tab type of tag \cr
[,4] \tab value \tab list \tab tag's details, depend on type of a tag \cr
}

There are a few types of tags listed below:
 TODO
}
\description{
Download the list of all tags with extra information abour judgments
}

