% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_dump_enrichments.R
\name{get_dump_enrichments}
\alias{get_dump_enrichments}
\title{Get all enrichments tags (EXPERIMENTAL)}
\usage{
get_dump_enrichments()
}
\value{
List of enrichment tags as returned from API.
}
\description{
Get the list of all tags with extra information abour judgments
}

