% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_dump_judgments.R
\name{get_dump_judgments}
\alias{get_dump_judgments}
\title{Get all judgments}
\usage{
get_dump_judgments(start_date = NULL, end_date = NULL,
  modification_date = NULL, simplify = FALSE)
}
\arguments{
\item{start_date}{Any date/time object that could be properly converted with
\code{as.POSIXct}, or a string in a format "\%Y-\%m-\%d".
Represents the earliest allowed judgment's date on the list.
If missing, no time limit is set.}

\item{end_date}{Any date/time object that could be properly converted with
\code{as.POSIXct}, or a string in a format "\%Y-\%m-\%d".
Represents the latest allowed judgment's date on the list.
If missing, no time limit is set.}

\item{modification_date}{Any date/time object that could be properly converted with
\code{as.POSIXct}, or a string in format "\%Y-\%m-\%dT\%H:\%M:\%S".
Allows to select judgments which were modified later than the specified
time. If missing, no time limit is set.}

\item{simplify}{Logical. If \code{TRUE} results will be returned as
\code{data.frame}.}
}
\value{
The list of judgments as returned from API.
}
\description{
Get the list of all judgments within (and including) given dates.
}
\examples{
\dontrun{
full <- get_dump_judgments()
lastchanges <- get_dump_judgments(modification_date = Sys.Date() - 7)
}
\donttest{
# Download judgments from last month
lastmonth <- get_dump_judgments(start_date = Sys.Date() - 30,
                               end_date = Sys.Date())
 }
}

