% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_scChambers.R
\name{get_dump_scChambers}
\alias{get_dump_scChambers}
\title{Download all Supreme Court chambers}
\usage{
get_dump_scChambers(flatten = FALSE)
}
\arguments{
\item{flatten}{logical, whether flatten information about divisions of every
 chamber}
}
\value{
If \code{flatten = FALSE} (default) data.frame as described in
  \code{\link[saos]{scchambers}}.
If \code{flatten = TRUE} data.frame with rows corresponding to divisions
  and five columns: information about chamber (\code{id, name}) and
  information about division (\code{div_id, div_name, div_fullName}).
}
\description{
Download information about all chambers of the Supreme Court of Poland.
}
\seealso{
\code{\link[saos]{scchambers}}, \code{\link[saos]{get_courts}}
}

