% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_judgments}
\alias{get_judgments}
\title{Get judgments with given IDs}
\usage{
get_judgments(idlist)
}
\arguments{
\item{idlist}{integer vector with judgments' IDs}
}
\value{
data.frame with rows corresponding to judgments, or NULL if none
 judgment is available
}
\description{
Get all judgments, if available, with IDs from a given vector
}
\examples{
get_judgments(128334)
get_judgments(c(128334, 77354))
}

