% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_judgments}
\alias{get_judgments}
\title{Get judgments with given IDs}
\usage{
get_judgments(idlist)
}
\arguments{
\item{idlist}{integer vector with judgments' IDs}
}
\value{
data.frame with rows corresponding to judgments, or NA if none
 judgment is available
}
\description{
Get all judgments, if available, with IDs from a given vector
}
\examples{
\donttest{
# single ID
get_judgments(128334)

# vector of IDs
get_judgments(c(128334, 77354))

# vector of IDs with non-existent judgment
get_judgments(c(128334, 1, 77354))
 }
}

