% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_judgments_dump.R
\name{get_judgments_dump}
\alias{get_judgments_dump}
\title{Get all judgments}
\usage{
get_judgments_dump(start_date = NULL, end_date = NULL)
}
\arguments{
\item{start_date}{An object of class \code{Date} or character in format
"YYYY-MM-DD". Represents the earliest allowed judgment's date on the list.
If missing, no time limit is set.}

\item{end_date}{An object of class \code{Date} or character in format
  "YYYY-MM-DD". Represents the latest allowed judgment's date on the list.
  If missing, no time limit is set.}
}
\value{
List of judgments as returned from API.
}
\description{
Get the list of all judgments within (and including) given dates.
}
\examples{
\donttest{
dump <- get_judgments_dump(start_date = "2015-01-15", end_date = "2015-01-20")
 }
}

