% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_scChambers}
\alias{get_scChambers}
\title{Download all Supreme Court chambers}
\usage{
get_scChambers(flatten = FALSE)
}
\arguments{
\item{flatten}{logical, whether flatten information about divisions of every
 chamber}
}
\value{
data.frame with rows corresponding to chambers or, if flatten == TRUE,
 to their divisions
}
\description{
Download information about all chambers of the Supreme Court of Poland.
}
\seealso{
\code{\link[saos]{get_courts}}
}

