% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/open_judgments.R
\name{open_judgments}
\alias{open_judgments}
\alias{open_judgments.numeric}
\alias{open_judgments.saos_search}
\title{Open SAOS judgments' pages}
\usage{
open_judgments(x, limit = 5)

\method{open_judgments}{numeric}(x, limit = 5)

\method{open_judgments}{saos_search}(x, limit = 5)
}
\arguments{
\item{x}{An object used to select a method.}

\item{limit}{Integer. Limit the number of opened pages.}
}
\value{
(Invisibly) Vector of character strings with URLs of abstracts opened.
}
\description{
Open in web browser SAOS pages for given judgments.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Open in web browser SAOS pages for judgments with
given IDs.

\item \code{saos_search}: Open in web browser SAOS pages for judgments with
from the given search results.
}}
\examples{
\donttest{

open_judgments(1:3)

judgments <- search_judgments(limit = 5, verbose = FALSE)
open_judgments(judgments)
 }
}

