% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search_judgments}
\alias{search_judgments}
\title{Search for judgments}
\usage{
search_judgments(query = NULL, limit = 200, force = FALSE)
}
\arguments{
\item{query}{Search pattern as a named list; pairs name-value are treated
as query parameters with corresponding values, see Details.}

\item{limit}{Limit the number of search results.}

\item{force}{If TRUE, force search request even if it seems extreme.}
}
\value{
data.frame with rows corresponding to judgments, or NULL if none
 judgment is available
}
\description{
Search for judgments matching the given query
}
\details{
The available query parameters:
}
\examples{
\dontrun{
search_judgments(list(dateFrom = "10-11-2014"))
 }
}

