% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search_judgments}
\alias{search_judgments}
\title{Search for judgments}
\usage{
search_judgments(all = NULL, legalBase = NULL,
  referencedRegulation = NULL, keyword = NULL, courtName = NULL,
  judgeName = NULL, judgmentDateFrom = NULL, judgmentDateTo = NULL,
  limit = 200, force = FALSE)
}
\arguments{
\item{all}{search everywhere for given phrase}

\item{legalBase}{legal basis of judgments}

\item{referencedRegulation}{regulations referenced in judgments}

\item{keyword}{keywords of judgments}

\item{courtName}{name of the trial court}

\item{judgeName}{name of any judge}

\item{judgmentDateFrom}{judgments from and including given date,
format "YYYY-MM-DD"}

\item{judgmentDateTo}{judgments to and including given date,
format "YYYY-MM-DD"}

\item{limit}{Limit the number of search results.}

\item{force}{If TRUE, force search request even if it seems extreme.}
}
\value{
list of judgments
}
\description{
Search for judgments matching the given query
}
\examples{
\donttest{
search_judgments(judgmentDateFrom = "2014-11-20")
search_judgments(judgeName="Maria Tyszel", judgmentDateTo="2014-06-30")

# search with no query, various limit options
search_judgments()
search_judgments(limit = 50)
search_judgments(limit = 300)
search_judgments(limit = 300, force = TRUE)
 }
}

