% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search_judgments}
\alias{search_judgments}
\title{Search for judgments}
\usage{
search_judgments(all = NULL, legalBase = NULL,
  referencedRegulation = NULL, keyword = NULL, courtName = NULL,
  judgeName = NULL, judgmentDateFrom = NULL, judgmentDateTo = NULL,
  limit = 200, force = FALSE)
}
\arguments{
\item{all}{search everywhere for given phrase}

\item{legalBase}{legal basis of judgments}

\item{referencedRegulation}{regulations referenced in judgments}

\item{keyword}{keywords of judgments}

\item{courtName}{name of the trial court}

\item{judgeName}{name of any judge}

\item{judgmentDateFrom}{judgments from and including given date,
format "YYYY-MM-DD"}

\item{judgmentDateTo}{judgments to and including given date,
format "YYYY-MM-DD"}

\item{limit}{Limit the number of search results.}

\item{force}{If TRUE, force search request even if it seems extreme.}
}
\value{
data.frame with rows corresponding to judgments, or \code{NULL} if
 none judgment is available, and following columns (* means that column is
 a list with elements of given type, if impossible to unlist)
\tabular{rlll}{
n \tab name \tab type \tab description \cr
[,1] \tab id \tab integer \tab judgment ID in the repository \cr
[,2] \tab href \tab character \tab link to judgment's page in API \cr
[,3] \tab courtCases \tab character \tab case signature, indicates division
  of the trial court, a case number and a year of the judgment \cr
[,4] \tab judgmentType \tab character \tab one from "DECISION", "RESOLUTION",
  "SENTENCE", "REGULATION", "REASONS" \cr
[,5] \tab judges \tab data.frame* \tab two columns with judges names ("name")
  and their role ("specialRoles") \cr
[,6] \tab textContent \tab character \tab shortened version of full text of
  the judgment \cr
[,7] \tab keywords \tab character* \tab keywords associated with the given
  judgment \cr
[,8] \tab division \tab data.frame* \tab information about the court (columns
  "court.name", "court.code", "court.href") and its division (columns "name",
  "code", "href"), see \link{get_courts} for details \cr
[,9] \tab judgmentDate \tab character \tab date of the judgment in format
  YYYY-MM-DD \cr
}
}
\description{
Search for judgments matching the given query
}
\examples{
\dontrun{
search_judgments(judgmentDateFrom = "2014-11-20")
search_judgments(judgeName="Maria Tyszel", judgmentDateTo="2014-06-30")

# search with no query, various limit options
search_judgments()
search_judgments(limit = 50)
search_judgments(limit = 300)
search_judgments(limit = 300, force = TRUE)
 }
}

